requirements.txt
google-cloud-vision
google-cloud-storage
Pillow
pypdfium2
pytesseract
opencv-python
fastapi
pydantic
